feat: :sparkles: set the next review date of newly-created projects (#43)
Set the next review date of a newly-created project to <review interval>
days/weeks/months/years after today.

If we don't do this, the newly-created project will inherit the review
date from the template, which is probably far in the past (i.e. whenever
the template was created), and the newly-created project will
immediately appear in the user's list of projects to review.  This is
probably not what the user wants, since they just created the project
and will presumably review it after creation.

Signed-off-by: Joshua Berry <des@condordes.net>